# add index to semantic table for better perf