--- a/packages/server/src/main/resources/db/migration/V19__semantics-index.sql
+++ b/packages/server/src/main/resources/db/migration/V19__semantics-index.sql
@@ -1,0 +1,1 @@
+CREATE INDEX IF NOT EXISTS semantic_1 ON semantic (model_configuration_id, dtype);