### Summary
Add compound index to match hibernate query patterns. Metric below from local db with 45k rows.

Before
```
terarium=# explain analyze select p1_0.model_configuration_id,p1_0.id,p1_0.created_on,p1_0.source,p1_0.type,p1_0.updated_on,p1_0.distribution,p1_0.is_default,p1_0.reference_id from semantic p1_0 where p1_0.model_configuration_id='d5b82001-dafe-42a1-af40-8e5ad05bc453' and p1_0.dtype='ParameterSemantic'                                                                                                                                                    ;
                                                             QUERY PLAN
-------------------------------------------------------------------------------------------------------------------------------------
 Seq Scan on semantic p1_0  (cost=0.00..1747.53 rows=3 width=139) (actual time=2.727..2.727 rows=0 loops=1)
   Filter: ((model_configuration_id = 'd5b82001-dafe-42a1-af40-8e5ad05bc453'::uuid) AND ((dtype)::text = 'ParameterSemantic'::text))
   Rows Removed by Filter: 46769
 Planning Time: 0.051 ms
 Execution Time: 2.740 ms
(5 rows)
```

After
```
terarium=# explain analyze select p1_0.model_configuration_id,p1_0.id,p1_0.created_on,p1_0.source,p1_0.type,p1_0.updated_on,p1_0.distribution,p1_0.is_default,p1_0.reference_id from semantic p1_0 where p1_0.model_configuration_id='d5b82001-dafe-42a1-af40-8e5ad05bc453' and p1_0.dtype='ParameterSemantic';
                                                                  QUERY PLAN
-----------------------------------------------------------------------------------------------------------------------------------------------
 Bitmap Heap Scan on semantic p1_0  (cost=4.31..12.08 rows=2 width=139) (actual time=0.035..0.036 rows=0 loops=1)
   Recheck Cond: ((model_configuration_id = 'd5b82001-dafe-42a1-af40-8e5ad05bc453'::uuid) AND ((dtype)::text = 'ParameterSemantic'::text))
   ->  Bitmap Index Scan on semantic_1  (cost=0.00..4.31 rows=2 width=0) (actual time=0.034..0.034 rows=0 loops=1)
         Index Cond: ((model_configuration_id = 'd5b82001-dafe-42a1-af40-8e5ad05bc453'::uuid) AND ((dtype)::text = 'ParameterSemantic'::text))
 Planning Time: 0.208 ms
 Execution Time: 0.052 ms
(6 rows)

```